Oops: ctrl+c didn't stop the program
In a go-routine there is no simple
way of aggregating errors, so if
the context is cancelled (by pressing
ctrl+c) we carried on as if nothing
happened.

When all executables have finished,
check the context if it's been cancelled
and return it as an error.

Main will recognize that the context
has been cancelled and set the exit
code 128+(signal raised).

Test plan:
* Have a template with Timeout=5
  and an executable $(sleep 10).
  Press ctrl+c before the timeout
  elapses. Verify that ain exits
  with exit status 130 (128+2).
* Let the timeout elapse and
  verify error-message that executable
  timed out after 5 seconds.