Bug: two or more env-vars on the same line
Ain incorrectly made a line with two env-vars
into two separate lines.

Change strategy: aggregate all env-vars
and their values on a line first into
a temp-struct.

Then iterate over that temp-struct and
splice in the values into the original
line.

This handles the case where a ${VAR}
contains the actual value ${VAR} - so
we're not infinitely looping on the same
replacement.

Test plan:
* Have a template with an one
  line ${VAR}. Make sure VAR is
  undefined. Verify error on the ${VAR}
  line.
* Have a template with one line
  and ${VAR1} ${VAR2}. Set both to
  undefined. Verify error on
  both variables.
* Set ${VAR1} to some value and
  verify error on the second variable.
* Have a template with no variables
  and verify lines are passed through
  (i e no regression).
* Have a template with ${VAR1}.
  Set VAR1='${VAR1}'$'\n''$(ff)'
  and run the template. Verify via
  the expanded context that the '${VAR1}'
  is not expanded again (and again).
* Have a template with ${VAR1} ${VAR2}.
  Set VAR1="[Host]\nlocalhost\n[Backend]\ncurl\n"
  VAR2="[Headers]\nContent-Type: ugh\n"
  in the .env file.
  Verify that the resulting ain -p
  call handles newlines in variables still.