Refactor: backendInput creates its own damn file
Step three in making BackendInput more
self-sufficient, let it take care of
dumping in the body-content into a
temp file.

Also make it in charge of cleaning up
any tempfile. In doing this error-handling
becomes easier as it's clear when there are
errors with the backend-call or when it's
because of the temp-file creation or deletion.

Test plan:
* Add a unit test to provide the logic
  to drive the temp-file creation.
  Set the Body to empty and verify
  no file created. Call RemoveBodyTempFile()
  and verify nothing happens.
* Set the Body: []string{"some", "stuff"}.
  Call CreateTempBodyFile() and verify
  a ain-body* is created in /tmp.
  Verify contents is as in Body:
  Call RemoveBodyTempFile(false) and
  verify file removed.
* As above but set the PrintCommand
  flag to true and verify file created
  in the cwd instead of /tmp.
* As above but set the flag
  LeaveTempFile: true
  and call RemoveTempFile(false).
  Verify file left in cwd.
* As above but call
  RemoveTempFile(true). Verify
  file removed even though
  LeaveTempFile is true.