Bug: report temp-file error before exiting
If there was an error removing the body temp
file when pressing ctrl+c the error would be
swallowed, since we exit before printing.

Fix this by always printing the error
message before checking for a signal
raised.

Check the context for a raised signal
and if present return the temp file
error.

Test plan:
* Have a template with a [Body]
  section and fake a slow
  client (e g curl with --limit-rate 1).
  Hack in an error when removing the
  temp file (such as removing a file
  that doesn't exist).
  Press ctrl+c and verify that an
  error on removing the temp-file
  is reported before the program exits.