Add tokenization of the template lines
Add a tokenization step that takes the input line
and spits out one of five token-types (envvar, executable,
text, comment or error).

The tokenization runs through the line and scans
for $( ${ and # respectively. This way
escaping of all the special tokens can be
handled in one place.

Tokenization handles quoting and rewrites escaped
special symbols as text without the backtick.

We run Tokenize() three times, once when creating the sectionedTemplate,
then two times more for parsing envvars and for excutables respectively.
When parsing envvars, any text containing an envvar is no longer valid (since
this phase is not recursive), thus the tokenizer can be configured
to parse symbols as text.

When it's time to set the sections we extract the content from the tokens
(and it should only be text-tokens left).

If the tokenization fails we're aggregating fatals as these fatals
deal with valid lines with invalid tokens.

Test plan:
* Have a standard template (e g generated with ain -b).
  Insert an envvar tokenization error (${unclosed )
  and one executable tokenization error ($(unclosed )
  on separate lines.
  Verify fatal-message prints missing bracket and
  parenthesis respectively.
* Remove the unclosed executable and change to ${}.
  Verify error-message on empty variable.
* Set to ${ENV} but supply no value for ENV.
  Verify error-message on no value.
* Set VAR= with the template above.
  Verify error-message on empty value for
  ${VAR}.
* Set VAR='${VAR}' and verify no error-message
  and that this value is passed verbatim to the output
  of ain (no escaping needed),
  (use for example in a [Header] section).
* Set VAR='$(unclosed' and verify error-message
  with expanded context on unclosed executable.
* Set VAR="one\ntwo\n$(unclosed\nthree" in the .env
  file. Verify error message on unclosed executable
  on the correct line with the expanded context.
* Set VAR='`$(exec) # $(echo yo)'
  and verify that the executable
  is escaped and passed to the output and the comment
  applied.
* Escape the comment in the example above and
  verify it's also passed to the output.
* Escape the escaped comment by adding
  \ and verify the comment is once again
  a comment and the line ends with `.
* Add a $(nonexistent) executable in the template.
  Set VAR='$(nonexistent2)'. Verify error-messages
  on executable not found.
* Insert a ${printf '1\n\2\n ${VAR} $(ignored # comment'} into the
  template, verify the output sans the comment is passed
  verbatim (and no error on unclosed executable)
  to the output.
* Escape the comment above and verify it's passed
  verbatim too.
* Reproduce the error in issue #25 and verify this
  patch solves the issue.