Ooops: executable token rest of string
When returning a fatal from the tokenize
and the current token is executable,
the error message is the rest of the string
since we're at the end.

Returning idx+1 would cause an
null-byte to be written. This is
because the input is a long string,
split over \n and the indexed via
a slice of runes.

The capacity for the underlying
array (the original string), is almost
always larger than the slice of
runes over the individual rows.

Thus a null byte is written when
we go over length on the []rune
slice instead of panic:ing.

Test plan:
* Have an executable with an
  unterminated quote.
  Run
  go run cmd/ain/main.go template | xxd.
  Verify that no null-byte is
  printed with the error-message.
* Do the same but with an unclosed
  parenthesis.