Refactor: Add splitTextOnComment
$(cmd ${ENVVAR}) fails b/c the
previous tokenization did not
support nested tokens.

So the $() was consumed and the
${ENV} inside it was treated as
plain text.

This is a regression as ${ENV}s
are expanded before any executables.

Change the tokenization strategy
to tokenize per line, but
keep the expandedSourceMarker
as text (content and comment
respectively).

This is the first step, introducing
a split on a comment in a line.

Test plan:
* Tests pass.