Back to contents and comments per line
Introduce a new old way of having the
expanded template lines as text, but
this time split 'em on content and
comments so we can ignore the comments
when parsing.

Add a new way of iterating that takes
a tokenizer that splits the line whichever
way the caller wants to (i e env-vars or
executables), iterates through the tokens
and calls an interator if the token is anything
but text.

Then split the returned contents on newlines
and comments and insert new expandedTemplateLines.

This change the behavior so we no longer touch
tokens on a line where an envvar inserts a comment
before that token.

E g "${VAR1} ${VAR2}" where VAR1 = 'Yo #'.

The resulting line will be
Yo # ${VAR2}
since the comment cuts off the line
we don't need to replace the VAR2
since it will never be read.

Test plan:
* Test-cases added for the tricky part
  of splitting lines on newlines
  and separating comments from contents.