Remove []tokens, use content and comments
Iterate over the template in an old
new way (old as it was before, line by line).

But new as in we transition to tokens during
each phase (envvars and executables) - then the
result is stitched back together as content
and comments.

Also introduce a fatalContent for expanded
fatals - we want to keep the escaped
envvars and executables in the error-message
even tough they are removed in the content
string itself.

Do do this we keep a fatalContent and shadow
what happens with the tokens, all of this
to keep the `. This might be reworked
with indices instead since fatals are
throwaway when the template is assembled.

Test plan:
* Added some tests/ templates.
  The templates contains expected
  output as comments. Run the
  files and verify output corresponds
  to what's in the file.