Add first stab at e2e tests
Regression is a thing.

Setup is a runnable
template with yaml as
comments last in the file.

Yaml because it's a
fairly readable format and
commented out because then
the file is self-containing
and can act as examples.

Currently passes in:
env-vars, args
and then compares the
stdout, stderr and exitcode.

Test harness builds a binary
then runs the tests.
All hooked into go test.

Test plan:
* go test ./... includes
  e2e tests, which passes.